Try to fix issues with exposing enums.
