+++ src/ApollonEditor.ts
@@ -4,92 +4,7 @@ import { Store } from 'redux';
 import { Application } from './scenes/Application';
 import { State } from './components/Store';
 import { DiagramType } from './domain/Diagram';
-import { ElementKind } from './domain/Element';
-import { RelationshipKind } from './domain/Relationship';
-import { Styles } from './components/Theme';
-
-export interface UMLModel {
-  version: string;
-  type: DiagramType;
-  size: { width: number; height: number };
-  interactive: Selection;
-  assessments: { [id: string]: Assessment };
-  elements: { [id: string]: UMLElement };
-  relationships: { [id: string]: UMLRelationship };
-}
-
-export interface Assessment {
-  score: number;
-  feedback?: string;
-}
-
-export { DiagramType };
-export { ElementKind as ElementType };
-export { RelationshipKind as RelationshipType };
-export { Styles };
-export { UMLClassifier, UMLClassAssociation } from './domain/plugins/ClassDiagram';
-
-interface UMLBase {
-  id: string;
-  name: string;
-  bounds: { x: number; y: number; width: number; height: number };
-}
-
-export interface UMLElement extends UMLBase {
-  owner: string | null;
-  type: ElementKind;
-}
-
-export enum Direction {
-  Up,
-  Right,
-  Down,
-  Left,
-}
-
-export interface UMLRelationship extends UMLBase {
-  type: RelationshipKind;
-  path: { x: number; y: number }[];
-  source: {
-    element: string;
-    direction: Direction;
-  };
-  target: {
-    element: string;
-    direction: Direction;
-  };
-}
-
-export interface Selection {
-  elements: string[];
-  relationships: string[];
-}
-
-export enum ApollonMode {
-  Modelling = 'Modelling',
-  Exporting = 'Exporting',
-  Assessment = 'Assessment',
-}
-
-export type ApollonOptions = {
-  type?: DiagramType;
-  mode?: ApollonMode;
-  readonly?: boolean;
-  model?: UMLModel;
-  theme?: Partial<Styles>;
-};
-
-export interface ExportOptions {
-  filter: string[];
-}
-
-export interface SVG {
-  svg: string;
-  size: {
-    width: number;
-    height: number;
-  };
-}
+import { ApollonOptions, Selection, UMLModel, ExportOptions, SVG } from '.';
 
 export class ApollonEditor {
   private application: RefObject<Application> = createRef();
