+++ src/domain/plugins/DiagramType.ts
@@ -1,6 +1,7 @@
-export enum DiagramType {
-  ClassDiagram = 'ClassDiagram',
-  ObjectDiagram = 'ObjectDiagram',
-  ActivityDiagram = 'ActivityDiagram',
-  UseCaseDiagram = 'UseCaseDiagram',
-}
+export type DiagramType = typeof DiagramType[keyof typeof DiagramType];
+export const DiagramType = Object.freeze({
+  ClassDiagram: 'ClassDiagram' as 'ClassDiagram',
+  ObjectDiagram: 'ObjectDiagram' as 'ObjectDiagram',
+  ActivityDiagram: 'ActivityDiagram' as 'ActivityDiagram',
+  UseCaseDiagram: 'UseCaseDiagram' as 'UseCaseDiagram',
+});
